Add files via upload
This GitHub repository contains a machine learning-based sentiment analysis project that was carried out on an Amazon review dataset. The goal of this project is to classify the sentiments of customer reviews on various musical instruments. The project utilized several machine learning algorithms such as Decision Tree, Logistic Regression, SVM, Random Forest, Naive Bayes, and K-Nearest Neighbors.

The code is written in Python and includes a detailed description of the data preprocessing techniques used in the project. Additionally, the code includes the implementation of Grid Search and Cross-Validation methods to find the optimal hyperparameters for the models. The accuracy of each model was evaluated using the Confusion Matrix and F1 Score metrics.